[mylyn] Simplify getting a TaskRepository for a git repo
First, if bugtracker.url is set in the config, use it unconditionally.
Second, simplify the alternate mechanism (find a task repository
matching the git upstream host) by not checking against the resolved
local host name. Calling InetAddress.getLocalHost() here is not such
a good idea since it may be extremely slow over a VPN or especially
if there is no Internet connection at all, and this gets called on the
UI thread.

Bug: 520386
Change-Id: Ia20c33e5e486581c437c5a5cb4aeb490e91dbb65
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>