Reproducable license feature version
We do not specify the version of the license feature, thereby always
taking the latest available license. This can be problematic, especially
on maintenance branches, since a build from the same commit may include
different licenses at different times.

As shown by Ed on the mailing list, use a fixed version number of the
license in the target instead. That way, the license reference is bound
to that version number and is under full control of the project and the
build.

Change-Id: Iee19578a9ce24826b4d4dabcd08daa86c9230611
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>