Fix git history page double refresh bug on repo switch
When the repository is switched inputSet(RevCommit) is being called,
which calls initAndStartRevWalk(...) which in turn sets currentRepo.

This causes a repo switched event to be triggered which the
ShowFirstParentOnlyPrefAction reacted to with a refresh of the history
page and that calls inputSet(RevCommit) again.

This commit changes ShowFirstParentOnlyPrefAction to not trigger
a refresh on getting a repo changed event.

Bug: 551582
Change-Id: I621030d383d48eba7f1ddef27840ce1107e1b7fc
Signed-off-by: Tim Neumann <Tim.Neumann@advantest.com>