Use Apache HTTP by default
Add a preference with UI to choose between the JDK built-in HTTP
handling and Apache HTTP. Set the default to Apache HTTP. In case
of trouble the user can switch back to using the Java standard HTTP
client.

The advantage of the Apache HTTP client is that it allows using
Basic authentication when HTTPS connections are to be made over
an HTTP proxy. OpenJDK forbids that by default since Java 8u111,[1]
and changing the JDK configuration to re-allow it is not obvious
at all and cannot be done automatically from within Eclipse.

[1] https://bugs.openjdk.java.net/browse/JDK-8210814

Bug: 549832
Change-Id: I71928b627d3af4dae95bdab63e019c75ab7e6a9e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>