[preferences] Apply defaults in the right preference store
EGit used an override of FieldEditor.getPreferenceStore() to supply
an alternate preference store than the normal one set on the
FieldEditorPreferencePage. However, FieldEditor sometimes bypassed
getPreferenceStore(), in particular storing an editor that was reset
to the default value. As a result, the default was applied in the
wrong preference store.

Fix this by overriding FieldEditor.setPreferenceStore() instead, which
guarantees that the editors do also save defaults in the correct store.

Bug: 552325
Change-Id: Iba3d80c13ef5e9a8bfbe98428b5c73342f28324d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>