[history] Apply cell editor in ref filter dialog on OK
A cell editor normally applies its value when it loses the focus.
But when the OK button is clicked, it doesn't get such an event, and
thus the value does not get applied.

Focus the OK button explicitly when it was clicked to force a focus
change before the dialog closes (and before the filters are saved).

Change-Id: I923a5e46ebd86922f3825aa00acfd2f76b4ed8d0
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>