Make sure the GarbageCollectCommand gets a FileRepository
JGit's GarbageCollectCommand works only for FileRepository and for
DfsRepository. With the new RepositoryHandle it would fail. Since
we cannot derive RepositoryHandle from FileRepository because of the
latter's constructor we have to give the GarbageCollectCommand access
to the underlying real repository. Encapsulate this in a utility
method in RepositoryUtil; that way, we don't have to open up access
to RepositoryHandle.getDelegate().

Bug: 553135
Change-Id: Ib8a1c557e5e250c5bfdfc888f816d18af6a521e1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>