[history] Avoid concurrent access to the GitHistoryWalk
RevWalk is not thread-safe, but there are several threads using a
GitHistoryWalk instance: GitHistoryJob, and then FormatJob and the
FindToolbarJob, and also the UI job. The latter three use the walk
via SWTCommit.parseBody() to get the raw bytes of a commit to be
able to work with the commit message. The headers have already been
parsed by the GitHistoryJob.

This separation was done to save memory, see commit bc785070. But
it can cause CorruptObjectExceptions, which occur more frequently
recently.

Use a different (new) RevWalk to read the commit body to avoid such
problems. This works because the headers have already been parsed.

Bug: 553289
Change-Id: I54cc43ed0666e9b68b3a26dd230bcff98e40a249
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>