[tests] Try to stabilize IndexDiffCacheTest
This test waits for a listener to be called. IndexDiff calculations
happen in the background. So when several files are added to the
workspace, several calculations may get scheduled, and the test may
end up not waiting long enough to pick up the expected change but
operate on an IndexDiff computed before the wanted file was added.

Run multiple workspace changes in a workspace operation with
IWorkspace.AVOID_UPDATE.

Change-Id: Ia58db7a5f9ad1cf7dedec387d8b4499a87c69fbc
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>