Show visible and overall staging count when filtering
When the file name filter in the staging view is active, show both the
visible and overall number of staging entries, even if they are the
same.

This makes it more explicit that a filter has been applied (since there
can be a short delay between the input of the filter text and the
refresh of the view, especially with other jobs working concurrently).

Bug:553053
Change-Id: Idd6ae15f4b6dd5c5fe3967eddc813fac00875e5a
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>