Staging View: fix title label updates
Since the label provider/decorator setup changes in commit b2e98cd3
the staging view title could flicker between undecorated and decorated
label. DecoratingStyledCellLabelProvider.getStyledText() has no
protection for decorations being pending, so it mustn't be used
anymore to determine the view's title.

Instead use a simple DecoratingLabelProvider. Use it to update a
ViewerLabel object on any LabelProviderChangedEvent, and update the
form title only if that ViewerLabel has a new image or text. The
label provider's updateLabel() method has similar protection for
pending decorations, and so the flickering can be avoided.

Also use this ViewerLabel to set the form title image; this gives
the staging view the correct image (for instance, the "Gerrit repo"
image for Gerrit-enabled repositories).

Bug: 558360
Change-Id: I69f83b21703959610208d06abda3b2aa65d6a555
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>