Filter jgit.benchmarks and jgit.coverage from import
Exclude jgit.benchmarks and jgit.coverage via nature filters during
the Oomph project import. The predicate solution is simpler to
understand than the alternative implementation via excludedPaths pattern
at the source locator.

Also eliminate the prompting of the variable value for
eclipse.target.platform by setting the value to none.

Provide a variable for prompting for the egit target platform which
defaults to the first value in the list of choices.

Also assign an ID to the working set task so that the pending changes in
Oomph's serialization will kick in the next time you edit with the
latest version of Oomph.  The goal is to make the serialization more
robust with respect to task additions/removals in combination with any
subsequent textual merging.

I would suggest going with this version for now, until you've resolved
how you want to automatically do a Maven build that would correct the
errors in the two excluded projects.

Change-Id: I08274260d664792c2f200fe17616a05367944f64
Signed-off-by: Ed Merks <ed.merks@gmail.com>