close submodules before deleting worktree
On Windows, deleting a repository containing submodules fails due to
file handles staying open. Submodules must be closed before the workTree
files are deleted.
My assumption is that otherwise the meta information about the submodule
is deleted and subRepo in RemoveCommand#closeSubmoduleRepositories is
null, so the submodule is not closed.

Bug: 408846
Change-Id: I76e030990021ef9c381760701f2ea9cd33de1768
Signed-off-by: Alexander Nittka <alex@nittka.de>