Disable "sign" button if no GpgSigner is present
JGit is moving the GPG signing support into a separate, optional
fragment. Although EGit will be set up such that this fragment is
included when EGit is installed, it's still possible that there is
no GpgSigner present at run-time.

Disable the button in that case, and give it a tooltip telling the
user that there is no signing support. Since this indicates a mis-
configuration, I've opted for disabling the button instead of hiding
it altogether.

Bug: 559106
Change-Id: Ifc477ef242b68d9f2b5ce367b493ed047b2eec00
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>