[history] Less flashy tags and branches in dark mode
Many of the tag and branch labels in the Git History view used very
bright colors, which doesn't look that well in dark mode.

Newly the rendering adapts to the table background, which changes
when the theme changes. If the table has a dark background use
darker colors, and if needed, use white for the text.

Bug: 536171
Change-Id: I0fdb9b5d6ecf0e16b221f559c33c7d6a649b401a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>