[history page] Key binding for the refresh action
Register the refresh action as the global refresh action so that
the configured key binding (F5 by default) works.

The tooltip on the toolbar button doesn't show the bound key, though.
The GenericHistoryView has its own refresh action, which is not
registered for the global retarget action. Doing so there might be
another way to fix this in org.eclipse.team.ui, but may have unknown
effects on other history pages but the EGit one.

Bug: 567600
Change-Id: I3a8cde955d6bf024db097df57c1465681ea0c7cc
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>